requirements.txt
#
# Galaxy has traditionally managed all of its own dependencies (in the form of
# precompiled Python eggs) for a few reasons:
#
# 1. We can ensure that Galaxy works with the versions of the dependencies that
#    we require.
# 2. We can precompile them for numerous platforms, meaning that developers or
#    system administrators attempting to use Galaxy do not need to wait while
#    dependencies compile/install, nor do they need to have compilers or
#    various development libraries installed on their system.
#
# However, you may want to manage these dependencies on your own. If so, this
# file can be used with pip to install all of the bare minimum dependencies for
# starting a Galaxy server. Use this in conjunction with the
# `GALAXY_ENABLE_EGGS=False` environment variable.
#
# Note that except in the case of Whoosh, all dependencies will be installed at
# their newest versions. Galaxy has not been tested with these and you may
# (likely?) run into problems! The Galaxy Team does not support using Galaxy
# with dependency versions other than those in eggs.ini, however, if you do
# find an incompatibility between Galaxy and the latest version, we would
# appreciate if you would investigate and provide a code fix for the issue. We
# are trying to move away from our strict version dependency policy and this
# would greatly help in our efforts.

Paste
PasteDeploy
docutils
wchartype
pexpect
amqp
repoze.lru
Routes
WebOb
SQLAlchemy
Cheetah
pycrypto
MarkupSafe
WebHelpers
anyjson
kombu
numpy
bx-python # bx must be installed after numpy!
pysam
Mako
PyYAML
mercurial
pytz
Babel
decorator
Tempita
sqlalchemy-migrate
Beaker
Whoosh<2.5
paramiko
ssh
Fabric
boto
simplejson
bioblend
Parsley
WebError
http://svgfig.googlecode.com/files/svgfig-1.1.6.tgz
pydicom==0.9.9
nibabel==1.3.0
